general fixes
incl. carousel
