--- dbConnection.js
@@ -1,20 +1,0 @@
-// import { MongoClient, ServerApiVersion } from "mongodb";
-
-
-
-// export default function dbConnection(){
-//   const uri =
-//     "mongodb+srv://<" +
-
-//     ">@cluster0.pnanr.mongodb.net/?retryWrites=true&w=majority";
-//   const client = new MongoClient(uri, {
-//     useNewUrlParser: true,
-//     useUnifiedTopology: true,
-//     serverApi: ServerApiVersion.v1,
-//   });
-//   client.connect((err) => {
-//     const collection = client.db("game-review-siteDB").collection("users");
-//     console.log(collection)
-//     client.close();
-//   });
-// }
