--- src/components/WelcomeItem.vue
@@ -1,84 +1,0 @@
-<template>
-  <div class="item">
-    <i>
-      <slot name="icon"></slot>
-    </i>
-    <div class="details">
-      <h3>
-        <slot name="heading"></slot>
-      </h3>
-      <slot></slot>
-    </div>
-  </div>
-</template>
-
-<style lang="scss">
-$color_1: var(--color-text);
-$color_2: var(--color-heading);
-
-.item {
-  margin-top: 2rem;
-  display: flex;
-}
-.details {
-  flex: 1;
-  margin-left: 1rem;
-}
-i {
-  display: flex;
-  place-items: center;
-  place-content: center;
-  width: 32px;
-  height: 32px;
-  color: $color_1;
-}
-h3 {
-  font-size: 1.2rem;
-  font-weight: 500;
-  margin-bottom: 0.4rem;
-  color: $color_2;
-}
-@media (min-width: 1024px) {
-  .item {
-    margin-top: 0;
-    padding: 0.4rem 0 1rem calc(var(--section-gap) / 2);
-    &:before {
-      content: ' ';
-      border-left: 1px solid var(--color-border);
-      position: absolute;
-      left: 0;
-      bottom: calc(50% + 25px);
-      height: calc(50% - 25px);
-    }
-    &:after {
-      content: ' ';
-      border-left: 1px solid var(--color-border);
-      position: absolute;
-      left: 0;
-      top: calc(50% + 25px);
-      height: calc(50% - 25px);
-    }
-    &:first-of-type {
-      &:before {
-        display: none;
-      }
-    }
-    &:last-of-type {
-      &:after {
-        display: none;
-      }
-    }
-  }
-  i {
-    top: calc(50% - 25px);
-    left: -26px;
-    position: absolute;
-    border: 1px solid var(--color-border);
-    background: var(--color-background);
-    border-radius: 8px;
-    width: 50px;
-    height: 50px;
-  }
-}
-
-</style>
