+++ src/views/DetailView.vue
@@ -5,15 +5,15 @@
         <div id="gameDetail">
           <div id="iconsContainer">
             <div id="starIcon">
-              <v-icon name="star" scale="1.2" color="#e50931" v-for="star in starAmount" :key="star"/>
+              <v-icon name="star" scale="1.2" color="#e50931" v-for="star in starAmount" :key="star" />
             </div>
             <div id="trailerIcon">
 
               <v-icon name="ticket-alt" scale="1.8" color="#e5e5e5" />
             </div>
           </div>
-<h4> {{game.ratings}}</h4>
-          <img :src="game.background_image" loading="lazy"/>
+          <h4> {{ game.ratings }}</h4>
+          <img :src="game.background_image" loading="lazy" />
         </div>
         <div id="descriptionsContainer">
           <div id="gameTitle">
@@ -61,13 +61,13 @@ export default {
 
     };
   },
-  props: {id: Number},
+  props: { id: Number },
   components: {
 
 
   },
   computed: {
-    starAmount(){
+    starAmount() {
       return Math.floor(Math.random() * 5) + 1
     }
   },
@@ -75,10 +75,10 @@ export default {
     this.getGameDetail()
   },
   methods: {
-    async getGameDetail(){
+    async getGameDetail() {
       this.showLoading = true;
       try {
-        const data  = await fetchGame(this.id);
+        const data = await fetchGame(this.id);
         console.log(data);
         this.game = data;
       } catch (error) {
@@ -120,10 +120,12 @@ $background-color_1: rgb(255, 196, 0);
   justify-content: center;
   align-items: center;
 }
+
 iframe {
   width: 100%;
   max-width: 650px;
 }
+
 #containerDetail {
   width: 100%;
   height: 100%;
@@ -132,9 +134,11 @@ iframe {
   justify-content: center;
   align-items: center;
 }
+
 #rateContainer {
   width: 100%;
 }
+
 #rateNumber {
   color: $color_1;
   font-size: 17px;
@@ -149,11 +153,13 @@ iframe {
   align-items: center;
   font-weight: bold;
 }
+
 .iconStyles {
   text-align: center;
   font-size: 14px;
   font-weight: bold;
 }
+
 #myList {
   background: #ff142d;
   border-radius: 25px;
@@ -170,26 +176,31 @@ iframe {
   align-items: center;
   transition: transform 0.5s;
   margin-bottom: 30px;
+
   &:hover {
     transform: scale(1.2);
     cursor: pointer;
     background: #c20b1d;
   }
 }
+
 #playIcon {
   margin-left: 10px;
 }
+
 #gameDetail {
   display: flex;
   flex-direction: column;
   width: 45%;
   justify-content: space-around;
+
   p {
     color: $color_1;
     font-family: $font-family_1;
     text-align: justify;
   }
 }
+
 #descriptionsContainer {
   height: 100%;
   width: 100%;
@@ -198,6 +209,7 @@ iframe {
   justify-content: space-between;
   align-items: center;
   padding-right: 20px;
+
   h1 {
     color: $color_2;
     text-shadow: black 0.1em 0.1em 0.2em;
@@ -206,6 +218,7 @@ iframe {
     text-align: center;
   }
 }
+
 #descriptions {
   border-radius: 20px;
   width: 90%;
@@ -213,9 +226,11 @@ iframe {
   display: grid;
   grid-template-columns: repeat(3, 1fr);
   grid-gap: 30px;
+
   span {
     color: $color_3;
   }
+
   p {
     color: $color_2;
     font-family: $font-family_3;
@@ -229,6 +244,7 @@ iframe {
     text-align: center;
   }
 }
+
 #iconsContainer {
   width: 100%;
   display: flex;
@@ -237,24 +253,29 @@ iframe {
   font-size: 14px;
   font-weight: bold;
   color: $color_1;
+
   p {
     margin-right: 10px;
     color: $color_1;
     font-weight: bold;
   }
 }
+
 #trailerIcon {
   display: flex;
   justify-content: center;
   align-items: center;
+
   p {
     text-align: center;
   }
 }
+
 #gameTitle {
   width: 58%;
   text-align: center;
 }
+
 #gameDescription {
   color: $color_1;
   font-family: $font-family_4;
@@ -268,6 +289,7 @@ iframe {
   text-align: center;
   width: 60%;
 }
+
 p {
   span {
     font-weight: 100;
@@ -275,11 +297,13 @@ p {
     padding-top: 10px;
   }
 }
+
 @media only screen and (max-width: 699px) {
   iframe {
     height: 200px;
   }
 }
+
 @media only screen and (max-width: 450px) {
   #descriptions {
     display: grid;
@@ -287,20 +311,22 @@ p {
     grid-gap: 30px;
   }
 }
+
 @media only screen and (max-width: 566px) {
   #iconsContainer {
     display: flex;
     flex-direction: column;
   }
+
   #starIcon {
     display: flex;
     width: 100%;
   }
+
   #trailerIcon {
     display: flex;
     width: 100%;
     justify-content: flex-end;
   }
 }
-
 </style>
