+++ src/views/GameView.vue
@@ -1,6 +1,6 @@
 <script setup lang="ts">
-import TheWelcome from '@/components/TheWelcome.vue'
-import GamesF from "@/components/TheWelcome.vue";
+import CarouselComponent from '@/components/CarouselComponent.vue'
+import GamesF from "@/components/CarouselComponent.vue";
 </script>
 <template>
 
@@ -8,13 +8,13 @@ import GamesF from "@/components/TheWelcome.vue";
   <div>
     <h4>All Available Games</h4>
 
-    <TheWelcome :key="data.id" :type="GamesF.data.type" v-for="genre in data"/> />
+    <CarouselComponent :key="data.id" :type="GamesF.data.type" />
   </div>
 </template>
 <script lang="ts">
 
 export default {
-  name: "games",
+  name: "Games",
   components: {
 
   },
@@ -31,18 +31,11 @@ export default {
 </script>
 
 <style>
- #app h4 {
+#app h4 {
   color: #e9e9e9;
   text-align: center;
-  font-family: Arial, Helvetica, sans-serif;
   padding-top: 30px;
   padding-bottom: 30px;
-}
-
-@media only screen and (max-width: 599px) {
-  #app h4 {
-    font-size: 14px;
-  }
 }
 </style>
