--- src/views/SignUp.vue
@@ -1,9 +1,0 @@
-<script setup lang="ts">
-import SignUp from '@/components/SignUpPage.vue'
-</script>
-
-<template>
-    <main>
-        <SignUp />
-    </main>
-</template>
