--- src/views/TestView.vue
@@ -1,9 +1,0 @@
-<script setup lang="ts">
-import TestPage from '@/components/TestPage.vue'
-</script>
-
-<template>
-    <main>
-        <TestPage />
-    </main>
-</template>
